feat(controller): add `RedirectView` and flash attributes to the methods of FolderController

+++ src/main/java/com/file/storage/controller/FolderController.java
@@ -3,10 +3,18 @@ package com.file.storage.controller;
 import com.file.storage.dto.folder.FolderDeleteRequest;
 import com.file.storage.dto.folder.FolderRenameRequest;
 import com.file.storage.dto.folder.FolderUploadRequest;
+import com.file.storage.exception.folder.InvalidFolderDeleteRequestException;
+import com.file.storage.exception.folder.InvalidFolderRenameRequestException;
+import com.file.storage.exception.folder.InvalidFolderUploadRequestException;
 import com.file.storage.service.FolderService;
+import com.file.storage.util.ValidationUtils;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 
 @Controller
 @RequestMapping("/folders")
@@ -16,20 +24,47 @@ public class FolderController {
     private final FolderService folderService;
 
     @PostMapping
-    public String uploadFolder(@ModelAttribute("folderUploadRequest") FolderUploadRequest folderUploadRequest) {
+    public RedirectView uploadFolder(@Valid @ModelAttribute("folderUploadRequest") FolderUploadRequest folderUploadRequest,
+                                     BindingResult bindingResult,
+                                     RedirectAttributes redirectAttributes) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFolderUploadRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
         folderService.uploadFolder(folderUploadRequest);
-        return "redirect:/";
+
+        redirectAttributes.addFlashAttribute("success", "Folder uploaded successfully");
+        return new RedirectView("/");
     }
 
     @PutMapping
-    public String renameFolder(@ModelAttribute("folderRenameRequest") FolderRenameRequest folderRenameRequest) {
+    public RedirectView renameFolder(@Valid @ModelAttribute("folderRenameRequest") FolderRenameRequest folderRenameRequest,
+                                     BindingResult bindingResult,
+                                     RedirectAttributes redirectAttributes) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFolderRenameRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
         folderService.renameFolder(folderRenameRequest);
-        return "redirect:/";
+
+        redirectAttributes.addFlashAttribute("success", "FOlder renamed successfully");
+        return new RedirectView("/");
     }
 
     @DeleteMapping
-    public String deleteFolder(@ModelAttribute("folderDeleteRequest") FolderDeleteRequest folderDeleteRequest) {
+    public RedirectView deleteFolder(@Valid @ModelAttribute("folderDeleteRequest") FolderDeleteRequest folderDeleteRequest,
+                                     BindingResult bindingResult,
+                                     RedirectAttributes redirectAttributes) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFolderDeleteRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
         folderService.deleteFolder(folderDeleteRequest);
-        return "redirect:/";
+
+        redirectAttributes.addFlashAttribute("success", "Folder deleted successfully");
+        return new RedirectView("/");
     }
 }
